+++ actix-web/CHANGES.md
@@ -5,10 +5,12 @@
 ### Added
 
 - Add `Resource::{get, post, etc...}` methods for more concisely adding routes that don't need additional guards.
+- Add `Compress::with_predicate()` method for customizing when compression is applied.
 
 ### Changed
 
 - Handler functions can now receive up to 16 extractor parameters.
+- The `Compress` no longer compresses image or video content by default.
 
 ## 4.3.1 - 2023-02-26